+++ src/deluge/gui/ui/rename/rename_clip_ui.cpp
@@ -15,7 +15,7 @@
  * If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "gui/ui/rename/rename_clipname_ui.h"
+#include "gui/ui/rename/rename_clip_ui.h"
 #include "definitions_cxx.hpp"
 #include "extern.h"
 #include "gui/l10n/l10n.h"
@@ -26,19 +26,19 @@
 #include "model/output.h"
 #include "model/song/song.h"
 
-RenameClipNameUI renameClipNameUI{};
+RenameClipUI renameClipUI{};
 
-RenameClipNameUI::RenameClipNameUI() {
+RenameClipUI::RenameClipUI() {
 	title = "Clip Name";
 }
 
-bool RenameClipNameUI::opened() {
+bool RenameClipUI::opened() {
 	bool success = QwertyUI::opened();
 	if (!success) {
 		return false;
 	}
 
-	enteredText.set(&clip->clipName);
+	enteredText.set(&clip->name);
 
 	displayText();
 
@@ -47,12 +47,12 @@ bool RenameClipNameUI::opened() {
 	return true;
 }
 
-bool RenameClipNameUI::getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows) {
+bool RenameClipUI::getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows) {
 	*cols = 0b11;
 	return true;
 }
 
-ActionResult RenameClipNameUI::buttonAction(deluge::hid::Button b, bool on, bool inCardRoutine) {
+ActionResult RenameClipUI::buttonAction(deluge::hid::Button b, bool on, bool inCardRoutine) {
 	using namespace deluge::hid::button;
 
 	// Back button
@@ -82,53 +82,44 @@ ActionResult RenameClipNameUI::buttonAction(deluge::hid::Button b, bool on, bool
 	return ActionResult::DEALT_WITH;
 }
 
-void RenameClipNameUI::enterKeyPress() {
+void RenameClipUI::enterKeyPress() {
 
-	// Don't allow duplicate names on clips of a single output.
-	if (!clip->clipName.equalsCaseIrrespective(&enteredText)) {
+	// If actually changing it...
+	if (!clip->name.equalsCaseIrrespective(&enteredText)) {
+		// don't let user set a name that is a duplicate of another name that has been set for another clip
 		if (clip->output->getClipFromName(&enteredText)) {
 			display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_DUPLICATE_NAMES));
 			return;
 		}
 	}
-	clip->clipName.set(&enteredText);
+	clip->name.set(&enteredText);
 	exitUI();
 }
 
-bool RenameClipNameUI::exitUI() {
+bool RenameClipUI::exitUI() {
 	display->setNextTransitionDirection(-1);
 	close();
 	return true;
 }
 
-ActionResult RenameClipNameUI::padAction(int32_t x, int32_t y, int32_t on) {
-
-	// Audition pad
-	if (x == kDisplayWidth + 1) {
-		return instrumentClipView.padAction(x, y, on);
-	}
+ActionResult RenameClipUI::padAction(int32_t x, int32_t y, int32_t on) {
 
 	// Main pad
-	else if (x < kDisplayWidth) {
+	if (x < kDisplayWidth) {
 		return QwertyUI::padAction(x, y, on);
 	}
 
 	// Otherwise, exit
-	else {
-		if (on && !currentUIMode) {
-			if (sdRoutineLock) {
-				return ActionResult::REMIND_ME_OUTSIDE_CARD_ROUTINE;
-			}
-			exitUI();
+	if (on && !currentUIMode) {
+		if (sdRoutineLock) {
+			return ActionResult::REMIND_ME_OUTSIDE_CARD_ROUTINE;
 		}
+		exitUI();
 	}
 
 	return ActionResult::DEALT_WITH;
 }
 
-ActionResult RenameClipNameUI::verticalEncoderAction(int32_t offset, bool inCardRoutine) {
-	if (Buttons::isShiftButtonPressed() || Buttons::isButtonPressed(deluge::hid::button::X_ENC)) {
-		return ActionResult::DEALT_WITH;
-	}
-	return instrumentClipView.verticalEncoderAction(offset, inCardRoutine);
+ActionResult RenameClipUI::verticalEncoderAction(int32_t offset, bool inCardRoutine) {
+	return ActionResult::DEALT_WITH;
 }