+++ src/deluge/gui/ui/rename/rename_midi_cc_ui.cpp
@@ -15,30 +15,49 @@
  * If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "gui/ui/rename/rename_output_ui.h"
+#include "gui/ui/rename/rename_midi_cc_ui.h"
 #include "definitions_cxx.hpp"
 #include "extern.h"
 #include "gui/l10n/l10n.h"
-#include "gui/views/arranger_view.h"
+#include "gui/views/automation_view.h"
 #include "hid/buttons.h"
 #include "hid/display/display.h"
 #include "hid/led/pad_leds.h"
+#include "model/instrument/midi_instrument.h"
 #include "model/output.h"
 #include "model/song/song.h"
 
-RenameOutputUI renameOutputUI{};
+RenameMidiCCUI renameMidiCCUI{};
 
-RenameOutputUI::RenameOutputUI() {
-	title = "Track Name";
+RenameMidiCCUI::RenameMidiCCUI() {
+	title = "CC Name";
 }
 
-bool RenameOutputUI::opened() {
+bool RenameMidiCCUI::opened() {
 	bool success = QwertyUI::opened();
 	if (!success) {
 		return false;
 	}
 
-	enteredText.set(&output->name);
+	Clip* clip = getCurrentClip();
+
+	int32_t cc = clip->lastSelectedParamID;
+
+	// if we're not dealing with a real cc number
+	// then don't allow user to edit the name
+	if (cc < 0 || cc == CC_EXTERNAL_MOD_WHEEL || cc >= kNumRealCCNumbers) {
+		return false;
+	}
+
+	MIDIInstrument* midiInstrument = (MIDIInstrument*)clip->output;
+
+	std::string_view name = midiInstrument->getNameFromCC(cc);
+	if (!name.empty()) {
+		enteredText.set(name.data());
+	}
+	else {
+		enteredText.clear();
+	}
 
 	displayText();
 
@@ -47,12 +66,12 @@ bool RenameOutputUI::opened() {
 	return true;
 }
 
-bool RenameOutputUI::getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows) {
+bool RenameMidiCCUI::getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows) {
 	*cols = 0b11;
 	return true;
 }
 
-ActionResult RenameOutputUI::buttonAction(deluge::hid::Button b, bool on, bool inCardRoutine) {
+ActionResult RenameMidiCCUI::buttonAction(deluge::hid::Button b, bool on, bool inCardRoutine) {
 	using namespace deluge::hid::button;
 
 	// Back button
@@ -82,34 +101,25 @@ ActionResult RenameOutputUI::buttonAction(deluge::hid::Button b, bool on, bool i
 	return ActionResult::DEALT_WITH;
 }
 
-void RenameOutputUI::enterKeyPress() {
+void RenameMidiCCUI::enterKeyPress() {
 
-	if (enteredText.isEmpty()) {
-		return;
-	}
+	Clip* clip = getCurrentClip();
+	MIDIInstrument* midiInstrument = (MIDIInstrument*)clip->output;
+	int32_t cc = clip->lastSelectedParamID;
 
-	// If actually changing it...
-	if (!output->name.equalsCaseIrrespective(&enteredText)) {
-		// if this is an audio output
-		// don't let user set a name that is a duplicate of another name that has been set for another audio output
-		// Sean: do we only want to do this for audio output's?
-		if (currentSong->getAudioOutputFromName(&enteredText)) {
-			display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_DUPLICATE_NAMES));
-			return;
-		}
-	}
+	midiInstrument->setNameForCC(cc, enteredText.get());
+	midiInstrument->editedByUser = true; // need to set this to true so that the name gets saved with the song / preset
 
-	output->name.set(&enteredText);
 	exitUI();
 }
 
-bool RenameOutputUI::exitUI() {
+bool RenameMidiCCUI::exitUI() {
 	display->setNextTransitionDirection(-1);
 	close();
 	return true;
 }
 
-ActionResult RenameOutputUI::padAction(int32_t x, int32_t y, int32_t on) {
+ActionResult RenameMidiCCUI::padAction(int32_t x, int32_t y, int32_t on) {
 
 	// Main pad
 	if (x < kDisplayWidth) {
@@ -127,6 +137,6 @@ ActionResult RenameOutputUI::padAction(int32_t x, int32_t y, int32_t on) {
 	return ActionResult::DEALT_WITH;
 }
 
-ActionResult RenameOutputUI::verticalEncoderAction(int32_t offset, bool inCardRoutine) {
+ActionResult RenameMidiCCUI::verticalEncoderAction(int32_t offset, bool inCardRoutine) {
 	return ActionResult::DEALT_WITH;
 }